+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^1.1.3",
     "i18next": "^21.10.0",
     "i18next-http-backend": "^1.4.4",
+    "normalize.css": "^8.0.1",
     "react": "^18.2.0",
     "react-animations": "^1.0.0",
     "react-countup": "^6.3.2",
