+++ src/components/LossesList/LossesList.js
@@ -1,11 +1,7 @@
 import React from 'react';
 import useFetch from "../../services/useFetch";
-import i18n from "i18next";
-import {initReactI18next, useTranslation} from "react-i18next";
-import translationsEn from "../../locals/translationsEn";
-import translationsUk from "../../locals/translationsUk";
+import { useTranslation} from "react-i18next";
 import CountUp from "react-countup";
-import {CButtonGroup, CButton} from "@coreui/react";
 import '@coreui/coreui/dist/css/coreui.min.css';
 import './lossesList.scss';
 import people from '../../assets/icons/icon-people.svg';
@@ -24,21 +20,9 @@ import auto from '../../assets/icons/icon-auto.svg';
 import special from '../../assets/icons/icon-special.svg';
 import milLogo from "../../assets/images/mil-logo.svg";
 
-i18n.use(initReactI18next).init({
-    resources: {
-        en: {translation: translationsEn},
-        uk: {translation: translationsUk},
-    },
-    lng: 'uk',
-    fallbackLng: 'uk',
-    interpolation: {escapeValue: false}
-});
+
 const LossesList = () => {
     const {t} = useTranslation();
-    const onChange = (event) => {
-        event.preventDefault();
-        i18n.changeLanguage(event.target.value);
-    };
 
     const {
         data,
@@ -50,26 +34,23 @@ const LossesList = () => {
 
 
     return (<>
-            <div className="change__lang">
-                <select name="language" onChange={onChange}>
-                    <option type="" value="ua">{t('ukrainian')}</option>
-                    <option value="en">{t('english')}</option>
-                </select>
 
+            <div className="container">
+                <header className="App-header">
+                    <div className="header__wrapper">
+                        <p className="subtitle">{t('General')}</p>
+                        <h1 className="title">{t('Losses')}</h1>
+                        <div className="date">
+                            <div>{data?.data.date}</div>
+                            <div className="day">{data?.data.day} {t('Day')}</div>
+                        </div>
+                    </div>
+                    <a href="https://www.mil.gov.ua/"><img src={milLogo}
+                                                           alt=""/></a>
+                </header>
             </div>
 
-            <header className="App-header">
-                <div className="header__wrapper">
-                    <p className="subtitle">{t('General')}</p>
-                    <h1 className="title">{t('Losses')}</h1>
-                    <div className="date">
-                        <div>{data?.data.date}</div>
-                        <div className="day">{data?.data.day} {t('Day')}</div>
-                    </div>
-                </div>
-                <a href="https://www.mil.gov.ua/"><img src={milLogo}
-                                                       alt=""/></a>
-            </header>
+
             <div className="container">
                 <div className="losses__list">
                     <div className="losses__item">
@@ -80,7 +61,7 @@ const LossesList = () => {
                                      end={data?.data.stats.personnel_units}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
 
                             />
                         </div>
@@ -93,7 +74,7 @@ const LossesList = () => {
                                      end={data?.data.stats.artillery_systems}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -106,7 +87,7 @@ const LossesList = () => {
                                      end={data?.data.stats.planes}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}/>
+                                     duration={3}/>
                         </div>
                     </div>
                     <div className="losses__item">
@@ -117,7 +98,7 @@ const LossesList = () => {
                                      end={data?.data.stats.warships_cutters}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
 
                             />
                         </div>
@@ -130,7 +111,7 @@ const LossesList = () => {
                                      end={data?.data.stats.atgm_srbm_systems}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -142,7 +123,7 @@ const LossesList = () => {
                                      end={data?.data.stats.tanks}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -154,7 +135,7 @@ const LossesList = () => {
                                      end={data?.data.stats.mlrs}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -167,7 +148,7 @@ const LossesList = () => {
                                      end={data?.data.stats.helicopters}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -179,7 +160,7 @@ const LossesList = () => {
                                      end={data?.data.stats.uav_systems}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -191,7 +172,7 @@ const LossesList = () => {
                                      end={data?.data.stats.cruise_missiles}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -203,7 +184,7 @@ const LossesList = () => {
                                      end={data?.data.stats.armoured_fighting_vehicles}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -215,7 +196,7 @@ const LossesList = () => {
                                      end={data?.data.stats.aa_warfare_systems}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -227,7 +208,7 @@ const LossesList = () => {
                                      end={data?.data.stats.vehicles_fuel_tanks}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
@@ -239,7 +220,7 @@ const LossesList = () => {
                                      end={data?.data.stats.special_military_equip}
                                      scrollSpyOnce={true}
                                      enableScrollSpy={true}
-                                     duration={1}
+                                     duration={3}
                             />
                         </div>
                     </div>
