+++ src/components/LossesList/lossesList.scss
@@ -1,27 +1,31 @@
 .losses__list {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  grid-column-gap: 15px;
-  grid-row-gap: 1em;
-  justify-content: space-evenly;
-  padding-bottom: 50px;
+ display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+
+
 
   .losses__item {
+    margin: 20px 40px 20px 40px;
+    width: 286px;
+    height: 367px;
     border: 1px none;
-    border-radius: 25px;
+    border-radius: 31.2016px;
     display: flex;
-    width: 300px;
-    height: 180px;
-    background-color: rgba(255, 255, 255, 0.58);
-
-    box-shadow: rgba(50, 50, 93, 0.25) 0 50px 100px -20px, rgba(0, 0, 0, 0.3) 0px 30px 60px -30px, rgba(10, 37, 64, 0.35) 0px -2px 6px 0px inset;
-
+    flex-direction: column;
+    align-items: center;
+    background: linear-gradient(90.71deg, rgba(245, 245, 245, 0.67) 0%, rgb(255, 255, 255) 100%);
+    justify-content: space-around;
     &_img {
-      width: 40%;
+      width: 70%;
       margin: 15px;
     }
 
     .current__wrapper {
+      width: 255.71px;
+      height:135.74px;
+      border-radius: 0 0 31px 31px / 0px 0px 31px 31px;
+      background-color: rgba(255, 0, 0, 0);
       align-items: center;
       display: flex;
       flex-direction: column;
